Separate net-utils from install command
